patch local_connect_and_auth in python rather than the file
