+++ tests/test_pyspark_gateway.py
@@ -7,6 +7,8 @@ import psutil
 import requests
 import pandas
 
+import pyspark
+
 from pyspark_gateway import server
 from pyspark_gateway import PysparkGateway
